Removed reference.qmd file
File was a holdover from when the structure of the reference guide was a bookdown. This was causing issues when rendering, even though it wasn't being called in the yaml file. Deleted information within document to see if any issues came out, and nothing adverse happened. Rather, website rendered successfully. So deleting file, but I have a back-up saved in case we need or want it later.